backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx>=0.27.0
openai==1.30.1
anthropic==0.26.0
google-genai>=1.0.0
groq==0.8.0
asyncio-throttle==1.0.2
python-multipart==0.0.9
